day6/backend/requirements.txt
amqp==5.3.1
aniso8601==10.0.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
async-timeout==5.0.1
billiard==4.2.2
blinker==1.9.0
cachelib==0.13.0
celery==5.5.3
cffi==2.0.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Caching==2.3.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security==5.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
idna==3.11
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
prompt_toolkit==3.0.52
pycparser==2.23
python-dateutil==2.9.0.post0
pytz==2025.2
redis==7.0.1
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
WTForms==3.2.1
